--- src/test/java/dandi/dandi/auth/application/JwtTokenParserTest.java
@@ -1,36 +1,0 @@
-package dandi.dandi.auth.application;
-
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-
-import dandi.dandi.auth.exception.UnauthorizedException;
-import java.util.Map;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-
-class JwtTokenParserTest {
-
-    private final JwtTokenDecoder jwtTokenDecoder = new JwtTokenDecoder();
-
-    @DisplayName("token을 받아 header를 반환한다.")
-    @Test
-    void parseHeaders() {
-        String token = "eyJraWQiOiI4NkQ4OEtmIiwiYWxnIjoiUlMyNTYifQ.eyJpc3MiOiJodHRwczovL2FwcGxlaWQuYXBwbGUuY29tIiwiYXVkIjoiY29tLndoaXRlcGFlay5zZXJ2aWNlcyIsImV4cCI6MTU5ODgwMDEyOCwiaWF0IjoxNTk4Nzk5NTI4LCJzdWIiOiIwMDAxNDguZjA2ZDgyMmNlMGIyNDgzYWFhOTdkMjczYjA5NzgzMjUuMTcxNyIsIm5vbmNlIjoiMjBCMjBELTBTOC0xSzgiLCJjX2hhc2giOiJ1aFFiV0gzQUFWdEc1OUw4eEpTMldRIiwiZW1haWwiOiJpNzlmaWl0OWIzQHByaXZhdGVyZWxheS5hcHBsZWlkLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjoidHJ1ZSIsImlzX3ByaXZhdGVfZW1haWwiOiJ0cnVlIiwiYXV0aF90aW1lIjoxNTk4Nzk5NTI4LCJub25jZV9zdXBwb3J0ZWQiOnRydWV9.GQBCUHza0yttOfpQ-J5OvyZoGe5Zny8pI06sKVDIJaQY3bdiphllg1_pHMtPUp7FLv3ccthcmqmZn7NWVoIPkc9-_8squ_fp9F68XM-UsERKVzBvVR92TwQuKOPFr4lRn-2FlBzN4NegicMS-IV8Ad3AKTIRMIhvAXG4UgNxgPAuCpHwCwEAJijljfUfnRYO-_ywgTcF26szluBz9w0Y1nn_IIVCUzAwYiEMdLo53NoyJmWYFWu8pxmXRpunbMHl5nvFpf9nK-OGtMJrmZ4DlpTc2Gv64Zs2bwHDEvOyQ1WiRUB6_FWRH5FV10JSsccMlm6iOByOLYd03RRH2uYtFw";
-
-        Map<String, String> headers = jwtTokenDecoder.parseHeaders(token);
-
-        assertThat(headers).containsKeys("alg", "kid");
-    }
-
-    @DisplayName("유효하지 않은 token의 header를 추출하려고 하면 예외를 발생시킨다.")
-    @Test
-    void parseHeaders_InvalidToken() {
-        String invalidToken = "invalidToken";
-
-        assertThatThrownBy(() -> jwtTokenDecoder.parseHeaders(invalidToken))
-                .isInstanceOf(UnauthorizedException.class)
-                .hasMessage("유효하지 않은 토큰입니다.");
-    }
-}
